Update repositories logic

+++ src/items/service.py
@@ -1,7 +1,7 @@
 from dataclasses import dataclass
+from uuid import UUID
 
 from src.repositories.base.items import ItemsRepository
-from src.schemas import PyObjectId
 
 from .schemas import Item, ItemCreate, ItemUpdate
 
@@ -11,32 +11,45 @@ class ItemsService:
     repository: ItemsRepository
 
     async def add(self, item_create: ItemCreate) -> Item:
-        inserted_id = await self.repository.add(item_create.model_dump())
-
-        inserted_item = await self.repository.get(inserted_id)
-        assert inserted_item is not None, "Item not found after insert"
-        return Item(**inserted_item)
+        inserted_item = await self.repository.add(
+            name=item_create.name, price=item_create.price, count=item_create.count
+        )
+        return inserted_item
 
     async def get_all(self, count: int, offset: int) -> list[Item]:
         result = await self.repository.get_many(count, offset)
-        return [Item(**item) for item in result]
+        return result
 
-    async def get(self, id: PyObjectId) -> Item | None:
+    async def get(self, id: UUID) -> Item | None:
         result = await self.repository.get(id)
         if result:
-            return Item(**result)
+            return result
         return None
 
     async def update(
         self,
-        id: PyObjectId,
+        id: UUID,
         values: ItemUpdate,
     ) -> Item | None:
-        new_values = values.model_dump(exclude_defaults=True)
-        if new_values:
-            updated_item = await self.repository.update(id, new_values)
-            return Item(**updated_item)
-        return None
+        item = await self.repository.get(id)
+        if item is None:
+            return None
 
-    async def delete(self, id: PyObjectId) -> int:
+        if values.count is not None:
+            item.count = values.count
+
+        if values.name is not None:
+            item.name = values.name
+
+        if values.price is not None:
+            item.price = values.price
+
+        await self.repository.update(id, item)
+        return item
+
+    async def delete(self, id: UUID) -> bool:
+        item = await self.repository.get(id)
+        if not item:
+            return False
+
         return await self.repository.delete(id)
